Ajustes

+++ src/infra/database/services/tenants/tenancy-connection.service.ts
@@ -1,5 +1,5 @@
 import { configTenancyDefault, configDefault } from '../../../../../ormconfig';
-import { Injectable } from '@nestjs/common';
+import { Injectable, OnModuleInit } from '@nestjs/common';
 
 import { TenantEntity } from '../../entities/tenants/tenant.entity';
 import { MysqlConnectionOptions } from 'typeorm/driver/mysql/MysqlConnectionOptions';
@@ -7,30 +7,48 @@ import { IDataSource } from 'src/helpers/class/data-source.interface';
 import { ITenancyConnection } from 'src/domain/repositories/tenants/tenancy-connection';
 
 @Injectable()
-export class TenancyConnection extends ITenancyConnection {
+export class TenancyConnection extends ITenancyConnection implements OnModuleInit {
     constructor() {
+        console.log(' aqyi')
         super();
-
-        this.createConnectionDefault();
     }
-
+    connectionFailedCount = 0;
     tenancy: IDataSource;
     tenanciesConnections: {
         name: TenantEntity['name'];
         dataSource: IDataSource;
     }[] = [];
 
+    async onModuleInit() {
+        console.log(' aqyi2')
+        // this.createConnectionDefault();
+        this.createConnectionsTenancies();
+    }
+
     async createConnectionsTenancies(): Promise<void> {
         const connectionDefault = await this.getConnectionDefault();
 
-        connectionDefault.getRepository(TenantEntity)
+        connectionDefault
+            .getRepository(TenantEntity)
             .find({
                 where: {
                     // version: "V1"
-                }
-
+                },
             })
             .then(async (tenants) => {
+                if (tenants.length === 0) {
+                    await connectionDefault.getRepository(TenantEntity).save({
+                        name: 'base_tenanty_1',
+                        version: 'V1',
+                        host: configTenancyDefault['host'],
+                        port: configTenancyDefault['port'],
+                        username: configTenancyDefault['username'],
+                        database: 'base_tenanty_1',
+                        password: configTenancyDefault['password'],
+                    })
+                    this.createConnectionsTenancies();
+                    return
+                }
                 for (let index = 0; index < tenants.length; index++) {
                     const tenancy = tenants[index];
 
@@ -68,10 +86,8 @@ export class TenancyConnection extends ITenancyConnection {
             .then((tenant) => tenant.name);
     }
 
-
     async getConnectionDefault(): Promise<IDataSource> {
-
-        if(!this.tenancy.isInitialized) {
+        if (!this.tenancy?.isInitialized) {
             await this.createConnectionDefault();
         }
 
@@ -79,34 +95,47 @@ export class TenancyConnection extends ITenancyConnection {
     }
 
     async createConnectionDefault(): Promise<void> {
-        this.tenancy = new IDataSource({
-            ...configTenancyDefault,
-        });
 
-        await this.tenancy.initialize()
-        .catch(async (error) => {
-            console.log('ERRO MYSQL',error)
+        if(this.connectionFailedCount > 0) {
+           await setTimeout(() => {return}, 2000)
+        }
+        if (this.connectionFailedCount < 8) {
+            this.connectionFailedCount++;
 
-            if(error.code.includes('ER_BAD_DB_ERROR')) {
-                let { database, ...configsDatabase } = configTenancyDefault;
-                let db = new IDataSource({
-                    ...configsDatabase,
-                } as MysqlConnectionOptions );
-                console.log('CRIANDO DATABASE')
+                this.tenancy = new IDataSource({
+                    ...configTenancyDefault,
+                });
 
-                await db.initialize();
-                await db.query(`CREATE DATABASE IF NOT EXISTS ${database}`).then(async () => {
-                    console.log(' DATABASE CRIADA')
-                    await db.destroy();
-                    this.createConnectionDefault();
-                }).catch((error) => {
-                    db.destroy();
+                await this.tenancy.initialize().catch(async (error) => {
+                    console.log('ERRO MYSQL', error);
 
-                    console.log(' DATABASE não Criada', error)
-                })
-                //aqui eu quero criar a database
+                    if (error.code.includes('ER_BAD_DB_ERROR')) {
+                        let { database, ...configsDatabase } = configTenancyDefault;
+                        let db = new IDataSource({
+                            ...configsDatabase,
+                        } as MysqlConnectionOptions);
+                        console.log('CRIANDO DATABASE');
 
-            }
-        });
+                        await db.initialize();
+                        await db
+                            .query(`CREATE DATABASE IF NOT EXISTS ${database}`)
+                            .then(async () => {
+                                console.log(' DATABASE CRIADA');
+                                await db.destroy();
+                                this.createConnectionDefault();
+                            })
+                            .catch((error) => {
+                                this.createConnectionDefault();
+                                db.destroy();
+
+                                console.log(' DATABASE não Criada', error);
+                            });
+                        //aqui eu quero criar a database
+                    }
+                });
+        } else {
+            console.log(' DATABASE CRIADA');
+            return;
+        }
     }
 }
